% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHRM_cov.R
\name{BHRM_cov}
\alias{BHRM_cov}
\title{Function that implement Bayesian Hierarchical Richard model with covariates}
\usage{
BHRM_cov(
  Y,
  t.values,
  X,
  seed.no = 1,
  burn = 2000,
  nmc = 2000,
  thin = 10,
  varrho = 0,
  pro.var.theta.2 = 2e-04,
  pro.var.theta.3 = 0.05,
  mu = 0,
  rho.sq = 1
)
}
\arguments{
\item{Y}{- N-by-T time sereis data for N countries for T days}

\item{t.values}{- N list for time perids for N countries}

\item{X}{- N-by-p design matrix (Should be column-wised standardized in advance)}

\item{burn}{- no of burn}

\item{nmc}{- no of iterations after burn}

\item{thin}{- no of thining for the nmc}

\item{varrho}{- hyper-parameter for the diffuse prior error variances ~ IG(varrho,varrho)}

\item{pro.var.theta.2}{- proposal variances for the MH algorithms to sample from theta.2.i}

\item{pro.var.theta.3}{- proposal variances for the MH algorithms to sample from theta.3.i}

\item{mu}{- mu}

\item{rho.sq}{- rho.sq}

\item{Index}{- i=1:N index for country || t=1:T_i index for time points (time ponts for each trajectory can differ) || j=1:p index for covariates}
}
\value{
A list of
       \item{thinned.theta.1.vec}{ - thinned.theta.1.vec}
       \item{thinned.theta.2.vec}{ - thinned.theta.2.vec}
       \item{thinned.theta.3.vec}{ - thinned.theta.3.vec}
       \item{thinned.xi.vec}{ - thinned.xi.vec}
       \item{thinned.alpha.1}{ - thinned.alpha.1}
       \item{thinned.alpha.2}{ - thinned.alpha.2}
       \item{thinned.alpha.3}{ - thinned.alpha.3}
       \item{thinned.beta.1.vec}{ - thinned.beta.1.vec}
       \item{thinned.beta.2.vec}{ - thinned.beta.2.vec}
       \item{thinned.beta.3.vec}{ - thinned.beta.3.vec}
       \item{thinned.sigma.sq}{ - thinned.sigma.sq}
       \item{thinned.sigma.1.sq}{ - thinned.sigma.1.sq}
       \item{thinned.sigma.2.sq}{ - thinned.sigma.2.sq}
       \item{thinned.sigma.3.sq}{ - thinned.sigma.3.sq}
       \item{mu}{ - mu}
       \item{rho.sq}{ - rho.sq}
}
\description{
Function that implement Bayesian Hierarchical Richard model with covariates
}
\examples{

}
