% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_sele.R
\name{credible_interval}
\alias{credible_interval}
\title{Function that do variable selection based on the mcmc samples of the coefficients}
\usage{
credible_interval(beta.vec, max.rank = 10, names = NULL)
}
\arguments{
\item{beta.vec}{- p-by-S matrix for S mcmc samples of p potential factors}

\item{max.rank}{- The number of top covariates selected based on the posterior mean}
}
\value{

}
\description{
Function that do variable selection based on the mcmc samples of the coefficients
}
\examples{

}
