% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\docType{data}
\name{design_matrix}
\alias{design_matrix}
\title{Potential covariates of 30 countries (noisy version)}
\format{
A data frame with 30 rows and 22 variables:
}
\source{
c(\url{https://data.worldbank.org/}, \url{https://apps.who.int/gho/data/node.main})
}
\usage{
design_matrix
}
\description{
A dataframe contains 20 potential factors for COVID-19 growth trajectory of 30 countries.
}
\keyword{datasets}
