% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Function that make predictions based on the trained model}
\usage{
predict(model, Y, i, num_days = 14, conf = 0.95, seed = 5)
}
\arguments{
\item{model}{- a list which is the output of function BHRM_cov, BHRM, or BRM}

\item{Y}{- N-by-T time sereis training data for N countries for T days which are used to get the model}

\item{i}{- index that indicates which subject will be predicted}

\item{num_days}{- integer that indicates the number of days to predict and the default value is 14}

\item{conf}{- a numerical value between 0 and 1 which indicates the confidence level of the interval}

\item{seed}{- random seed set in the function}
}
\value{

}
\description{
Function that make predictions based on the trained model
}
\examples{
data("time_series_data")
Y = time_series_data[, -c(1:2)]
seed.no = 1 ; burn = 20000 ; nmc = 20000 ; thin = 30; varrho = 0
pro.var.theta.2 = 0.0002 ; pro.var.theta.3 = 0.05; mu = 0 ; rho.sq = 1
res = BRM(Y = Y[1, ], seed.no = seed.no, burn = burn, nmc = nmc,
thin = thin, varrho = varrho, pro.var.theta.2 = pro.var.theta.2,
pro.var.theta.3 = pro.var.theta.3, mu = mu, rho.sq = rho.sq)
predict_list = predict(model, Y, 1)
}
