% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_sele.R
\name{var_sele}
\alias{var_sele}
\title{Function that do variable selection based on the mcmc samples of the coefficients}
\usage{
var_sele(beta.vec, max.rank = 10, names = NULL)
}
\arguments{
\item{beta.vec}{- p-by-S matrix for S mcmc samples of p potential factors}

\item{max.rank}{- The number of top covariates selected based on the posterior mean}
}
\value{

}
\description{
Function that do variable selection based on the mcmc samples of the coefficients
}
\examples{
data("design_matrix")
data("time_series_data")
Y = time_series_data[, -c(1:2)]; X = design_matrix[, -c(1:2)]
X = scale(X)
seed.no = 1 ; burn = 40000 ; nmc = 20000 ; thin = 30; varrho = 0
pro.var.theta.2 = 0.0002 ; pro.var.theta.3 = 0.05; mu = 0 ; rho.sq = 1
res_cov = BHRM_cov(Y = Y, X = X, seed.no = seed.no, burn = burn, nmc = nmc,
thin = thin, varrho = varrho, pro.var.theta.2 = pro.var.theta.2,
pro.var.theta.3 = pro.var.theta.3, mu = mu, rho.sq = rho.sq)
var_selection = var_sele(beta.vec = res_cov$thinned.theta.1.vec, names = names(X))
}
