% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_time_point.R
\name{flat_time_point}
\alias{flat_time_point}
\title{Function that calculates the flat time point of a Richard curve given the parameters}
\usage{
flat_time_point(theta.1, theta.2, theta.3, xi, gamma = 0.99)
}
\arguments{
\item{theta.1}{- parameter theta1 of Richard curve.}

\item{theta.2}{- parameter theta2 of Richard curve.}

\item{theta.3}{- parameter theta3 of Richard curve.}

\item{xi}{- parameter xi of Richard curve.}

\item{gamma}{- a progression constant between 0 and 1 which determines when it can be called flat time point.}
}
\value{
The flat time point of a Richard curve given the parameters.
}
\description{
Function that calculates the flat time point of a Richard curve given the parameters
}
\examples{
data("time_series_data")
Y = time_series_data[, -c(1:2)]
seed.no = 1 ; burn = 20000 ; nmc = 20000 ; thin = 30; varrho = 0
pro.var.theta.2 = 0.0002 ; pro.var.theta.3 = 0.05; mu = 0 ; rho.sq = 1
num_days = 14; t.values = c(1:(ncol(Y) - num_days))
Y = Y[, c(1:(ncol(Y) - num_days))]
res = BRM(Y = Y[1, ], t.values = t.values, seed.no = seed.no, burn = burn,
nmc = nmc, thin = thin, varrho = varrho, pro.var.theta.2 = pro.var.theta.2,
pro.var.theta.3 = pro.var.theta.3, mu = mu, rho.sq = rho.sq)
theta.1 = mean(res$thinned.theta.1.vec)
theta.2 = mean(res$thinned.theta.2.vec)
theta.3 = mean(res$thinned.theta.3.vec)
xi = mean(res$thinned.xi.vec)
flat_time_point(theta.1,theta.2,theta.3,xi)
}
