% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plot_RM}
\alias{plot_RM}
\title{Function that make a plot and can see a comparison between the observations and predictions}
\usage{
plot_RM(
  pred,
  y,
  upper = NULL,
  lower = NULL,
  num_days = 14,
  title = NULL,
  y_name = NULL,
  size_axis_text = 20,
  size_axis_title = 25,
  size_plot_title = 35
)
}
\arguments{
\item{pred}{- a vector of the mean of predicted values.}

\item{y}{- a vector of the observations}

\item{upper}{- the upper bound of the predicted values. If it's not specified, the plot won't draw the interval.}

\item{lower}{- the lower bound of the predicted values. If it's not specified, the plot won't draw the interval.}

\item{num_days}{- integer that indicates the number of days to predict and the default value is 14.}

\item{title}{- the title of the plot. If it's not specified, the plot won't include title.}

\item{y_name}{- the name of the y axis. If it's not specified, the name for the y axis will be "Values".}

\item{size_axis_text}{- the text size of y and x axis and the default value is 20}

\item{size_axis_title}{- the text size of y and x title and the default value is 25}

\item{size_plot_title}{- the text size of plot title and the default value is 35}
}
\value{
A plot that can see a comparison between the observations and predictions.
}
\description{
Function that make a plot and can see a comparison between the observations and predictions
}
\examples{
data("time_series_data")
Y = time_series_data[, -c(1:2)]
seed.no = 1 ; burn = 20000 ; nmc = 20000 ; thin = 30; varrho = 0
pro.var.theta.2 = 0.0002 ; pro.var.theta.3 = 0.05; mu = 0 ; rho.sq = 1
res = BRM(Y = Y[1, ], seed.no = seed.no, burn = burn, nmc = nmc,
thin = thin, varrho = varrho, pro.var.theta.2 = pro.var.theta.2,
pro.var.theta.3 = pro.var.theta.3, mu = mu, rho.sq = rho.sq)
predict_list = predict(model, Y, 1)
plot_RM(predict_list$prediction, Y[1, ], predict_list$upper, predict_list$lower)
}
